refactor: Docker development and production run scripts (#310)
* refactor: Deduplicate and improve api run logic

Combine run configurations for api into one script for improved clarity 
and better maintainability. Behaviour remains largely the same except 
that staging no longer runs with `—reload`, the default gunicorn workers
count goes from 2 to 1 (for dev and enterprise), there are no longer 
run commands to choose from for enterprise (you can still choose to 
override the default command though), and there are some additional 
configuration options available through environment variables.

* chore(devenv): Remove old uv project locations

* refactor(devenv): Deduplicate application running logic

API and worker now share a single start script as a base due to similar 
setup and then call their respective entrypoint scripts at the end for 
consistency and maintainability.

* feat(worker): Add run options from api script

* chore: Add temporary wrapper scripts

These scripts remain so that our Helm targets don’t break upon 
deployment.

+++ apps/worker/worker.sh
@@ -1,25 +1,52 @@
-#!/bin/bash
+#!/usr/bin/env bash
+set -euo pipefail
 
-if [ -n "$PROMETHEUS_MULTIPROC_DIR" ]; then
-    rm -r "$PROMETHEUS_MULTIPROC_DIR" 2> /dev/null
-    mkdir "$PROMETHEUS_MULTIPROC_DIR"
+# Default entrypoint for worker
+echo "Starting worker"
+
+# Script section to keep in sync with api.sh
+#### Start ####
+
+# Optional prefix and suffix for all python commands
+pre="${CODECOV_WRAPPER:-}"
+post="${CODECOV_WRAPPER_POST:-}"
+
+# Whether to ignore the prefix and suffix on migration commands
+if [[ -n "${CODECOV_WRAPPER_IGNORE_MIGRATE:-}" ]]; then
+  pre_migrate=""
+  post_migrate=""
+else
+  pre_migrate="$pre"
+  post_migrate="$post"
+fi
+
+# Berglas is used to manage secrets in GCP.
+berglas=""
+if [[ -f "/usr/local/bin/berglas" ]]; then
+  berglas="berglas exec --"
+fi
+
+#### End ####
+
+if [[ -n "${PROMETHEUS_MULTIPROC_DIR:-}" ]]; then
+  rm -r "${PROMETHEUS_MULTIPROC_DIR:?}"/* 2>/dev/null || true
+  mkdir -p "$PROMETHEUS_MULTIPROC_DIR"
 fi
 
 queues=""
-if [ "$CODECOV_WORKER_QUEUES" ]; then
+if [[ -n "${CODECOV_WORKER_QUEUES:-}" ]]; then
   queues="--queue $CODECOV_WORKER_QUEUES"
 fi
 
-if [ "$RUN_ENV" = "ENTERPRISE" ] || [ "$RUN_ENV" = "DEV" ]; then
-    python manage.py migrate
-    python manage.py migrate --database "timeseries"
-    python manage.py migrate --database "ta_timeseries"
-    python manage.py pgpartition --yes --skip-delete
+if [[ "${RUN_ENV:-}" = "ENTERPRISE" ]] || [[ "${RUN_ENV:-}" = "DEV" ]]; then
+  echo "Running migrations"
+  $pre_migrate $berglas python manage.py migrate $post_migrate
+  $pre_migrate $berglas python migrate_timeseries.py $post_migrate
+  $pre_migrate $berglas python manage.py pgpartition --yes --skip-delete $post_migrate
 fi
 
-if [ -z "$1" ];
-then
-  python main.py worker ${queues}
+if [[ -z "${1:-}" ]]; then
+  $pre $berglas python main.py worker $queues $post
 else
   exec "$@"
 fi
